% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.cdd}
\alias{climdex.cdd}
\title{Maximum Consecutive Dry Days}
\usage{
climdex.cdd(ci, spells.can.span.years = T, include.exact.dates = FALSE)
}
\arguments{
\item{include.exact.dates}{Logical, if TRUE, return a data frame with spell durations and the start and end dates of each spell; if FALSE, return only the spell durations.}
}
\description{
This function computes the climdex index CDD.
}
\details{
This function computes the climdex index CDD: the annual maximum length of dry spells, in days.
Dry spells are considered to be sequences of days where daily preciptation
is less than 1mm per day.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
