% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.csdi}
\alias{climdex.csdi}
\title{Cold Spell Duration Index}
\usage{
climdex.csdi(ci, spells.can.span.years = FALSE)
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{spells.can.span.years}{specifies whether spells can cross
year boundaries -- i.e., span years. The default for this is the same as
fclimdex.}
}
\description{
This function computes the climdex index CSDI.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index CSDI (Cold Spell Duration Index).

The cold spell duration index is defined as the number of days
each year which are part of a "cold spell". A "cold spell" is defined as a
sequence of 6 or more days in which the daily minimum temperature is below
the 10th percentile of daily minimum temperature for a 5-day running window
surrounding this day during the baseline period.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
