% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.prcptot}
\alias{climdex.prcptot}
\title{Total Daily Precipitation}
\usage{
climdex.prcptot(ci)
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation in wet days.
}
\description{
This function computes the climdex index PRCPTOT.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index PRCPTOT: the annual sum of precipitation in wet days
(days where precipitation is at least 1mm).
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
