% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.r20mm}
\alias{climdex.r20mm}
\title{Precipitation Exceeding 20mm Per Day}
\usage{
climdex.r20mm(ci)
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing the value of the index for each year.
}
\description{
This function computes the climdex index R20mm.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index R20mm: the annual count of days where daily precipitation is more than 20mm per day.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
