% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.r95ptot}
\alias{climdex.r95ptot}
\title{Total Daily Precipitation Exceeding 95\\\%ile Threshold}
\usage{
climdex.r95ptot(ci)
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation exceeding
the threshold.
}
\description{
This function computes the climdex index R95pTOT.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index R95pTOT: the annual sum of precipitation in days where daily precipitation exceeds the
95th percentile of daily precipitation in the base period.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
