% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.rx5day}
\alias{climdex.rx5day}
\title{Monthly Maximum 5-day Consecutive Precipitation}
\usage{
climdex.rx5day(
  ci,
  freq = c("monthly", "annual", "seasonal"),
  center.mean.on.last.day = FALSE,
  include.exact.dates = FALSE
)
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{freq}{Time frequency to aggregate to.}

\item{center.mean.on.last.day}{Whether to center the 5-day running mean on
the last day of the window, instead of the center day.}

\item{include.exact.dates}{Logical, if TRUE, return a data frame with the index values and exact dates for each month, season, or year; if FALSE, return only the index values.}
}
\value{
A vector containing the value of the index per time interval.
If \code{include.exact.dates} is TRUE, a data frame with additional exact dates is returned.
}
\description{
This function computes the climdex index Rx5day.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index Rx5day: monthly, seasonal or annual maximum 5-day consecutive precipitation.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
