% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.tnn}
\alias{climdex.tnn}
\title{Monthly Minimum of Daily Minimum Temperature}
\usage{
climdex.tnn(
  ci,
  freq = c("monthly", "annual", "seasonal"),
  include.exact.dates = FALSE
)
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{freq}{Time frequency to aggregate to.}

\item{include.exact.dates}{Logical, if TRUE, return a data frame with the index values and exact dates for each month, season, or year; if FALSE, return only the index values.}
}
\value{
A vector containing the value of the index per time interval.
If \code{include.exact.dates} is TRUE, a data frame with additional exact dates is returned.
}
\description{
This function computes the climdex index TNn.
}
\details{
This function takes a climdexInput object as input and computes
the monthly, seasonal or annual minimum of daily minimum temperature.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
