% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.tr}
\alias{climdex.tr}
\title{Tropical Nights}
\usage{
climdex.tr(ci)
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing the number of tropical nights for each year.
}
\description{
This function computes the climdex index TR.
}
\details{
This function takes a climdexInput object as input and computes the TR
(tropical nights) climdex index: that is, the annual count of days where
daily minimum temperature stays above 20 degrees Celsius.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
