% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_indices.R
\name{climdex.tx10p}
\alias{climdex.tx10p}
\title{Percent of Values Below 10th Percentile Daily Maximum Temperature}
\usage{
climdex.tx10p(ci, freq = c("monthly", "annual", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{freq}{Time frequency to aggregate to.}
}
\description{
This function computes the climdex index TX10p.
}
\details{
This function takes a climdexInput object as input and computes the
monthly, seasonal or annual percent of values below the 10th percentile of baseline
daily maximum temperature.
}
\references{
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
