% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexpcic_functions.R
\name{climdexInput.raw}
\alias{climdexInput.raw}
\title{Method for creating climdexInput object from vectors of data}
\usage{
climdexInput.raw(
  tmax = NULL,
  tmin = NULL,
  prec = NULL,
  tmax.dates = NULL,
  tmin.dates = NULL,
  prec.dates = NULL,
  base.range = c(1961, 1990),
  n = 5,
  northern.hemisphere = TRUE,
  tavg = NULL,
  tavg.dates = NULL,
  quantiles = NULL,
  temp.qtiles = c(0.1, 0.9),
  prec.qtiles = c(0.95, 0.99),
  max.missing.days = c(annual = 15, monthly = 3, seasonal = 6),
  min.base.data.fraction.present = 0.1
)
}
\arguments{
\item{tmax}{numeric vector containing the data on which the indices are to be computed. (degrees C)}

\item{tmin}{numeric vector containing the data on which the indices are to be computed. (degrees C)}

\item{prec}{numeric vector containing the data on which the indices are to be computed. (mm/day)}

\item{tmax.dates}{vector of type \code{PCICt} of the maximum date.}

\item{tmin.dates}{vector of type \code{PCICt} of the minimum date.}

\item{prec.dates}{vector of type \code{PCICt} of the precipitation date.}

\item{base.range}{a pair of 4 digit years which bound the data on which the base percentiles are calculated.}

\item{n}{the size of the window used when computing the percentiles used in \code{\link{climdex.tx10p}},
\code{\link{climdex.tn10p}}, \code{\link{climdex.tx90p}}, and \code{\link{climdex.tn90p}}.}

\item{northern.hemisphere}{whether the data came from
the northern hemisphere. If FALSE, data is assumed to have come from the
southern hemisphere. This is used when computing growing season length; if
the data is from the southern hemisphere, growing season length is the
growing season starting in the beginning of July of the year indicated,
running to the end of June of the following year.}

\item{tavg}{Average temperature, default \code{NULL}}

\item{tavg.dates}{Average temperature dates, default \code{NULL}}

\item{quantiles}{supply pre-computed quantiles. This is a list consisting of quantiles for each variable.}

\item{temp.qtiles}{modify the quantiles calculated. For example, specifying
temp.qtiles=c(0.10, 0.50, 0.90) would calculate the 10th, 50th, and 90th
percentiles for temperature.}

\item{prec.qtiles}{modify the quantiles calculated.}

\item{max.missing.days}{vector consisting of 'annual'
(the number of days that can be missing in a year), 'monthly' (the
number of days that can be missing in a month and 'seasonal' (the
number of days that can be missing in a season. If one month in a year fails
the test, the corresponding year will be omitted.}

\item{min.base.data.fraction.present}{TODO}

\item{min.base.fraction.present}{the minimum fraction
of data which must be present for a quantile to be calculated for a
particular day. If the fraction of data present is less than this threshold,
the quantile for that day will be set to NA.}
}
\value{
An object of class \code{\link{climdexInput-class}} for use with
other climdex methods.
}
\description{
This function creates a climdexInput object from data already ingested into
R.
}
\details{
This function takes input climate data at daily resolution, and produces as
output a ClimdexInput data structure. This data structure can then be passed
to any of the routines used to compute the Climdex indices. The indices
themselves are specified on the webpage cited in the references section.

For each temperature variable, there are separate lists of quantiles for
inbase and outbase, with these names. In both cases, quantiles within these
lists are named q10 for the 10th percentile and q90 for the 90th percentile.
Other percentiles would be named qnn for the nnth percentile. For the
outbase quantiles, each element in the list is a vector of length 365 (or 360
in the case of 360-day calendars), corresponding to one value for each day of
the year. For the inbase quantiles, each element in the list is an array of
dimensions \code{[365 or 360, nyr, nyr - 1]}, where nyr is the number of years in
the base period. Each value corresponds to a quantile for each day, for each
year, with a particular year replaced.

For precipitation variables, there is a named vector of quantiles, consisting
of at least q95 and q99.
}
\note{
Units are assumed to be mm/day for precipitation and degrees Celsius
for temperature. No units conversion is performed internally.
}
\examples{
library(PCICt)

## Create a climdexInput object from some data already loaded in and
## ready to go.

## Parse the dates into PCICt.
#tmax.dates <- PCICt::as.PCICt(do.call(paste, ec.1018935.tmax[,c("year",
#"jday")]), format="\%Y \%j", cal="gregorian")
#tmin.dates <- PCICt::as.PCICt(do.call(paste, ec.1018935.tmin[,c("year",
#"jday")]), format="\%Y \%j", cal="gregorian")
#prec.dates <- PCICt::as.PCICt(do.call(paste, ec.1018935.prec[,c("year",
#"jday")]), format="\%Y \%j", cal="gregorian")

## Load the data in.
#ci <- climdexInput.raw(ec.1018935.tmax$MAX_TEMP,
#ec.1018935.tmin$MIN_TEMP, ec.1018935.prec$ONE_DAY_PRECIPITATION,
#tmax.dates, tmin.dates, prec.dates, base.range=c(1971, 2000))
}
\references{
For Climdex Indices: \url{http://etccdi.pacificclimate.org/list_27_indices.shtml}
This function is from the \code{pacificclimate/climdex.pcic} repository.
The \code{climdex.pcic} package was developed and maintained by the Pacific Climate Impacts Consortium (PCIC).
For more information, visit the repository: \url{https://github.com/pacificclimate/climdex.pcic}
Special thanks to the contributors of the \code{climdex.pcic} package for their efforts in climate data analysis.
}
\keyword{climate}
\keyword{ts}
