% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_details.R
\name{.compress_spells_drop}
\alias{.compress_spells_drop}
\title{Compress Spells Drop}
\usage{
.compress_spells_drop(df, station_col)
}
\arguments{
\item{df}{A dataset containing climatic data.}

\item{station_col}{A character vector specifying the stations to analyse.}
}
\value{
compressed data frame
}
\description{
A helper function for \code{climatic_details} under the \code{hierarchical} option.
}
