% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instatClimatic-package.R
\docType{package}
\name{instatClimatic-package}
\alias{instatClimatic}
\alias{instatClimatic-package}
\title{instatClimatic: A set of climatic functions used in R-Instat}
\description{
This package contains a set of climatic functions used in R-Instat. This includes functions written specifically for R-Instat, as well functions written elsewhere that are modified for use in R-Instat.
}
\author{
\strong{Maintainer}: Lily Clements \email{lily@idems.international} (\href{https://orcid.org/0000-0001-8864-0552}{ORCID})

Authors:
\itemize{
  \item Danny Parsons
  \item Antoine Ntalumeso
  \item Stephen Lloyd
}

}
\keyword{internal}
