% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{na_check}
\alias{na_check}
\title{Check Missing Values Based on Conditions}
\usage{
na_check(
  x,
  na_type = c(),
  na_consecutive_n = NULL,
  na_max_n = NULL,
  na_max_prop = NULL,
  na_min_n = NULL,
  na_FUN = NULL,
  ...
)
}
\arguments{
\item{x}{A vector to check for missing values.}

\item{na_type}{A character vector specifying the types of checks to perform. Options include:
\itemize{
\item \code{"n"}: Total number of missing values (\verb{<= na_max_n}).
\item \code{"prop"}: Proportion of missing values (\verb{<= na_max_prop} in percentage).
\item \code{"n_non_miss"}: Minimum number of non-missing values (\verb{>= na_min_n}).
\item \code{"FUN"}: A custom function to evaluate missing values.
\item \code{"con"}: Maximum consecutive missing values (\verb{<= na_consecutive_n}).
}}

\item{na_consecutive_n}{Optional. Maximum allowed consecutive missing values.}

\item{na_max_n}{Optional. Maximum allowed missing values.}

\item{na_max_prop}{Optional. Maximum allowed proportion of missing values (in percentage).}

\item{na_min_n}{Optional. Minimum required non-missing values.}

\item{na_FUN}{Optional. A custom function to evaluate missing values.}

\item{...}{Additional arguments passed to the custom function \code{na_FUN}.}
}
\value{
Logical. Returns \code{TRUE} if all specified checks pass, otherwise \code{FALSE}.
}
\description{
Evaluates a vector against specified conditions for missing values.
}
