% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_sum}
\alias{summary_sum}
\title{Calculate Sum of Data}
\usage{
summary_sum(x, weights = NULL, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{Optional weights for the data.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The sum or weighted sum of the data.
}
\description{
Computes the sum or weighted sum of a dataset.
}
